[WGSL] webgpu:shader,validation,shader_io,group_and_binding:single_entry_point:* is failing https://bugs.webkit.org/show_bug.cgi?id=310664 rdar://150368198
Reviewed by Mike Wyrzykowski.

Move the validation that no two resources with the same group and binding are
used by a given entry point earlier into the pipeline as it should happen at
shader-creation time. Some of the logic can be reused in GlobalVariableRewriter,
but this patch doesn't yet take full advantage of it as that will require a
larger refactor.

* LayoutTests/http/tests/webgpu/webgpu/shader/validation/shader_io/group_and_binding-expected.txt:
* LayoutTests/platform/mac-wk2/TestExpectations:
* Source/WebGPU/WGSL/AttributeValidator.cpp:
(WGSL::AttributeValidator::validateIO): Deleted.
(WGSL::AttributeValidator::validateBuiltinIO): Deleted.
(WGSL::AttributeValidator::validateLocationIO): Deleted.
(WGSL::AttributeValidator::validateInterpolationIO): Deleted.
(WGSL::AttributeValidator::validateStructIO): Deleted.
(WGSL::validateIO): Deleted.
* Source/WebGPU/WGSL/AttributeValidator.h:
* Source/WebGPU/WGSL/CallGraph.h:
(WGSL::CallGraph::addGlobal const):
* Source/WebGPU/WGSL/GlobalVariableRewriter.cpp:
(WGSL::RewriteGlobalVariables::collectGlobals):
(WGSL::RewriteGlobalVariables::visitEntryPoint):
(WGSL::RewriteGlobalVariables::determineUsedGlobals):
(WGSL::RewriteGlobalVariables::insertStructs):
* Source/WebGPU/WGSL/IOValidator.cpp: Added.
(WGSL::IOValidator::IOValidator):
(WGSL::IOValidator::validateIO):
(WGSL::IOValidator::collectGlobals):
(WGSL::IOValidator::visit):
(WGSL::IOValidator::validateResources):
(WGSL::IOValidator::validateEntryPointIO):
(WGSL::IOValidator::validateBuiltinIO):
(WGSL::IOValidator::validateLocationIO):
(WGSL::IOValidator::validateInterpolationIO):
(WGSL::IOValidator::validateStructIO):
(WGSL::IOValidator::error):
(WGSL::validateIO):
* Source/WebGPU/WGSL/IOValidator.h: Copied from Source/WebGPU/WGSL/AttributeValidator.h.
* Source/WebGPU/WGSL/WGSL.cpp:
* Source/WebGPU/WebGPU.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/310359@main
