+++ Source/WebGPU/WGSL/IOValidator.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 Apple Inc. All rights reserved.
+ * Copyright (c) 2026 Apple Inc. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -31,6 +31,6 @@ namespace WGSL {
 
 class ShaderModule;
 
-std::optional<FailedCheck> validateAttributes(ShaderModule&);
+std::optional<FailedCheck> validateIO(ShaderModule&);
 
 } // namespace WGSL
